+++ breaditnow-customer-api/build.gradle
@@ -15,6 +15,9 @@ dependencies {
     // swagger
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.6.0'
 
+    // s3
+    implementation 'io.awspring.cloud:spring-cloud-starter-aws:2.4.4'
+
     // discord webhook
     implementation 'com.github.napstr:logback-discord-appender:1.0.0'
