+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/customer/presentation/CustomerController.java
@@ -2,12 +2,15 @@ package com.breaditnow.customer.domain.customer.presentation;
 
 import com.breaditnow.common.response.ApiSuccessResponse;
 import com.breaditnow.customer.domain.customer.application.CustomerInitializationService;
+import com.breaditnow.customer.domain.customer.application.CustomerProfileService;
+import com.breaditnow.customer.domain.customer.application.request.CustomerInfoUpdateRequest;
 import com.breaditnow.customer.domain.customer.application.request.CustomerInitRequest;
 import com.breaditnow.customer.domain.customer.application.response.CustomerInfoResponse;
 import com.breaditnow.customer.domain.customer.presentation.docs.CustomerControllerDocs;
 import com.breaditnow.customer.global.security.annotation.AuthCustomer;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 @RequestMapping("/api/v1/customer")
 @RestController
@@ -15,7 +18,7 @@ import org.springframework.web.bind.annotation.*;
 public class CustomerController  {
     private final CustomerInitializationService initializationService;
 //    private final CustomerInfoQueryService queryService;
-//    private final CustomerProfileService profileService;
+    private final CustomerProfileService profileService;
 //    private final PasswordVerifyQueryService securityService;
 //    private final NicknameDuplicateQueryService nicknameService;
 //    private final CustomerRegionService regionService;
@@ -32,7 +35,7 @@ public class CustomerController  {
     public ApiSuccessResponse<CustomerInfoResponse> getCustomerInfo(@AuthCustomer Long customerId) {
         return ApiSuccessResponse.of(queryService.getCustomerInfo(customerId));
     }
-
+   */
     // 내 정보 수정
     @PatchMapping(value = "/me/info", consumes = "multipart/form-data")
     public ApiSuccessResponse<CustomerInfoResponse> updateCustomerInfo(
@@ -42,6 +45,7 @@ public class CustomerController  {
         return ApiSuccessResponse.of(profileService.updateCustomerInfo(customerId, dto, profileImage));
     }
 
+    /*
     // 현재 비밀번호 확인
     @PostMapping("/me/verify-password")
     public ApiSuccessResponse<PasswordVerifyResponse> verifyPassword(@AuthCustomer Long customerId, @RequestBody PasswordVerifyRequest dto) {
