--- breaditnow-external/src/main/java/com/breaditnow/external/domain/s3/FileUploaderService.java
@@ -1,7 +1,0 @@
-package com.breaditnow.external.domain.s3;
-
-import org.springframework.web.multipart.MultipartFile;
-
-public interface FileUploaderService {
-	String upload(MultipartFile multipartFile, String dirName);
-}
