--- breaditnow-external/src/main/java/com/breaditnow/external/global/config/S3Config.java
@@ -1,34 +1,0 @@
-package com.breaditnow.external.global.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import com.amazonaws.auth.AWSCredentials;
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-
-@Configuration
-public class S3Config {
-	@Value("${aws.credentials.access-key}")
-	private String accessKey;
-
-	@Value("${aws.credentials.secret-key}")
-	private String secretKey;
-
-	@Value("${aws.region.static}")
-	private String region;
-
-	@Bean
-	public AmazonS3 amazonS3Client() {
-		AWSCredentials credentials = new BasicAWSCredentials(accessKey, secretKey);
-
-		return AmazonS3ClientBuilder
-			.standard()
-			.withCredentials(new AWSStaticCredentialsProvider(credentials))
-			.withRegion(region)
-			.build();
-	}
-}
